# services/data-service/openapi.json
{
  "openapi": "3.0.3",
  "info": { "title": "Data Service API", "version": "v1" },
  "servers": [{ "url": "/" }],
  "paths": {
    "/health": { "get": { "summary": "Health check", "responses": { "200": { "description": "OK" } } } },
    "/v1/prices": {
      "get": {
        "summary": "List prices",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "name": "company", "in": "query", "schema": {"type":"string"} },
          { "name": "skuId", "in": "query", "schema": {"type":"string"} },
          { "name": "from", "in": "query", "schema": {"type":"string", "example":"2025-01"} },
          { "name": "to", "in": "query", "schema": {"type":"string", "example":"2025-12"} }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/v1/forecast": {
      "get": { "summary": "List forecast", "security": [{ "ApiKeyAuth": [] }], "responses": { "200": { "description": "OK" } } }
    },
    "/v1/saleforecast": {
      "get": {
        "summary": "List sales forecast records",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "name": "anchor_month", "in": "query", "required": true, "schema": { "type": "string", "pattern": "^\\d{4}-(0[1-9]|1[0-2])$" } },
          { "name": "company_code", "in": "query", "schema": { "type": "string" } },
          { "name": "company_desc", "in": "query", "schema": { "type": "string" } },
          { "name": "material_code", "in": "query", "schema": { "type": "string" } },
          { "name": "material_desc", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "summary": "Create sales forecast record",
        "security": [{ "ApiKeyAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["anchor_month", "forecast_qty"],
                "properties": {
                  "anchor_month": { "type": "string", "pattern": "^\\d{4}-(0[1-9]|1[0-2])$" },
                  "company_code": { "type": "string" },
                  "company_desc": { "type": "string" },
                  "material_code": { "type": "string" },
                  "material_desc": { "type": "string" },
                  "forecast_qty": { "type": "number" },
                  "metadata": { "type": "object", "additionalProperties": true }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/v1/saleforecast/{recordId}": {
      "put": {
        "summary": "Update sales forecast record",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "name": "recordId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "anchor_month": { "type": "string", "pattern": "^\\d{4}-(0[1-9]|1[0-2])$" },
                  "company_code": { "type": "string" },
                  "company_desc": { "type": "string" },
                  "material_code": { "type": "string" },
                  "material_desc": { "type": "string" },
                  "forecast_qty": { "type": "number" },
                  "metadata": { "type": "object", "additionalProperties": true }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not Found" } }
      },
      "delete": {
        "summary": "Delete sales forecast record",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "name": "recordId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not Found" } }
      }
    }
    ,
    "/v1/monthly-access/seed-default": {
      "post": {
        "summary": "Seed default monthly access records",
        "security": [{ "ApiKeyAuth": [] }],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "anchor_months": {
                    "type": "array",
                    "items": { "type": "string", "pattern": "^\\d{4}-(0[1-9]|1[0-2])$" }
                  },
                  "users": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": ["user_email"],
                      "properties": {
                        "user_email": { "type": "string", "format": "email" },
                        "user_id": { "type": "string" },
                        "user_name": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" } }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  }
}